requirements.txt
numpy>=1.23
pandas>=1.5
PyYAML>=6.0
joblib>=1.3
tqdm>=4.66
tqdm-joblib>=0.0.3
psutil>=5.9
statsmodels>=0.14
matplotlib>=3.7
plotly>=5.18
kaleido>=0.2.1
